Time: 4 ms (48.11%) | Memory: 7 MB (16.46%) - LeetSync

+++ 389-find-the-difference/find-the-difference.cpp
@@ -3,7 +3,7 @@ public:
     char findTheDifference(string s, string t) {
         // 25.09.23 POTD
 
-        long long int d=0;
+        int d=0;
 
         for(int i=0; i<s.size(); i++){
             d -= s[i];
